+++ src/command_modules/azure-cli-interactive/azure/cli/command_modules/interactive/tests/latest/test_query_injection.py
@@ -7,8 +7,8 @@ import unittest
 import six
 import os
 import json
-from azclishell.util import parse_quotes
-from azclishell.gather_commands import GatherCommands
+from azure.cli.command_modules.interactive.azclishell.util import parse_quotes
+from azure.cli.command_modules.interactive.azclishell.gather_commands import GatherCommands
 
 
 def pass_gather(*_):
@@ -28,7 +28,7 @@ class QueryInjection(unittest.TestCase):
     """ tests using the query gesture for the interactive mode """
     def __init__(self, *args, **kwargs):
         super(QueryInjection, self).__init__(*args, **kwargs)
-        from azclishell.app import AzInteractiveShell
+        from azure.cli.command_modules.interactive.azclishell.app import AzInteractiveShell
         from azure.cli.testsdk import TestCli
         self.stream = six.StringIO()
         self.shell = AzInteractiveShell(TestCli(), output_custom=self.stream)