+++ src/command_modules/azure-cli-profile/azure/cli/command_modules/profile/_help.py
@@ -24,10 +24,10 @@ helps['login'] = """
             az login --service-principal -u http://azure-cli-2016-08-05-14-31-15 -p ~/mycertfile.pem --tenant contoso.onmicrosoft.com
         - name: Log in using a VM's system assigned identity
           text: >
-            az login --msi
+            az login --identity
         - name: Log in using a VM's user assigned identity. Client or object ids of the service identity also work
           text: >
-            az login --msi -u /subscriptions/<subscriptionId>/resourcegroups/myRG/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myID
+            az login --identity -u /subscriptions/<subscriptionId>/resourcegroups/myRG/providers/Microsoft.ManagedIdentity/userAssignedIdentities/myID
     """
 
 helps['account'] = """