+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -7,6 +7,7 @@ Release History
 ++++++
 * Support Autorest 3.0 based SDKs
 * warn the user that `az aks browse` won't work in Azure Cloud Shell
+* add `aks upgrade-connector` command to upgrade an existing connector
 
 2.0.27
 ++++++