+++ src/command_modules/azure-cli-acs/HISTORY.rst
@@ -8,6 +8,7 @@ Release History
 * Support Autorest 3.0 based SDKs
 * warn the user that `az aks browse` won't work in Azure Cloud Shell
 * add `aks upgrade-connector` command to upgrade an existing connector
+* `kubectl` config files are more readable block-style YAML
 
 2.0.27
 ++++++